Cleanup and flake8 fixes
